requirements.txt
slack_bolt==1.27.0
python-dotenv==1.0.1
google-genai==1.55.0